\name{rnaseqWrapper-package}
\alias{rnaseqWrapper-package}
\alias{rnaseqWrapper}
\docType{package}
\title{
A Wrapper for RNAseq analysis
}
\description{
This package wraps several aspects of RNAseq already implemented in R,
including DE analysis and analysis of variants.
For the development build, or to contribute changes to this package,
please see our repository at https://bitbucket.org/petersmp/rnaseqwrapper/
}
\details{
\tabular{ll}{
Package: \tab rnaseqWrapper\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-06\cr
License: \tab GPL\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mark Peterson

Maintainer: Mark Peterson <mark.phillip.peterson@gmail.com>

}
\references{
This package wraps access to several existing functions.
Look at each function to see what other pieces are used.
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
%~~ directory ~~
\keyword{ package }
\examples{
%~~ simple examples of the most important functions ~~
## See the examples in each function, for your interests

}
